tripsync/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.2
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-celery-beat==2.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-guardian==2.4.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-silk==0.5.3
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
flower==2.0.1
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jsonfield==3.1.0
kombu==5.5.2
MarkupSafe==3.0.2
msgpack==1.1.0
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
